MSVC: fix narrowing conversion warnings

+++ OgreMain/include/OgreGpuProgramParams.h
@@ -1501,8 +1501,8 @@ namespace Ogre {
         typedef vector<GpuSharedParametersUsage>::type GpuSharedParamUsageList;
 
         // Map that store subroutines associated with slots
-        typedef OGRE_HashMap<unsigned int, String> SubroutineMap;
-        typedef OGRE_HashMap<unsigned int, String>::const_iterator SubroutineIterator;
+        typedef OGRE_HashMap<size_t, String> SubroutineMap;
+        typedef OGRE_HashMap<size_t, String>::const_iterator SubroutineIterator;
 
     protected:
         SubroutineMap mSubroutineMap;
